Create lua test helper
